requirements.txt
streamlit
faiss-cpu
pillow
torch
transformers
tqdm
cohere